Create README.md
This printer.cfg file is current as of 4/22/23, and includes configuration information for the SKR Mini V3 3.0 main board. DO NOT USE THIS CONFIG UNLESS YOU HAVE INSTALLED THE SKR MINI V3 into your Anycubic Vyper. 

Additionally, I've included a precompiled firmware.bin file that can be copied to the root directory of a micro SD card and then used to flash the SKR Mini with Klipper.